:sparkles: feat : 프로그램 리스트 모델클래스 업데이트 및 인터벌 시간단위를 분에서 초로 수정

+++ frontend/lib/models/programs/program_list.dart
@@ -72,7 +72,7 @@ class ProgramItem {
 }
 
 class Program {
-  int? targetValue;
+  double? targetValue;
   IntervalInfo? intervalInfo;
 
   Program({this.targetValue, this.intervalInfo});
@@ -95,7 +95,7 @@ class Program {
 }
 
 class IntervalInfo {
-  int? duration;
+  double? duration;
   int? setCount;
   int? rangeCount;
   List<Ranges>? ranges;
@@ -129,8 +129,8 @@ class IntervalInfo {
 class Ranges {
   int? id;
   bool? isRunning;
-  int? time;
-  int? speed;
+  double? time;
+  double? speed;
 
   Ranges({this.id, this.isRunning, this.time, this.speed});
